% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gistic2ge.R
\name{base.summary}
\alias{base.summary}
\title{\code{base.summary} adds additional columns to DE summary table from gene expression data}
\usage{
base.summary(ge.mat, groups, groups.label = c("case", "control"),
  logged = TRUE)
}
\arguments{
\item{ge.mat}{gene expression matrix}

\item{groups}{sample indicator for groups}

\item{groups.label}{c(a, b) a = label for positive samples, b = label for negative samples}

\item{logged}{TRUE if gene expression is log2-transformed, base.summary provides statistics on unlogged data}
}
\description{
\code{base.summary} adds additional columns to DE summary table from gene expression data
}

