% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gistic2ge.R
\name{base.summary.wrapper}
\alias{base.summary.wrapper}
\title{\code{base.summary.wrapper} wrapper for \code{base.summary}}
\usage{
base.summary.wrapper(ge, gistic, logged = TRUE)
}
\arguments{
\item{ge}{internal ge object}

\item{gistic}{internal gistic object}

\item{logged}{TRUE if gene expression is log2-transformed, base.summary.wrapper provides statistics on unlogged data}
}
\description{
\code{base.summary.wrapper} wrapper for \code{base.summary}
}

