% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iEDGE_util.R
\name{calc_sobel}
\alias{calc_sobel}
\title{Performs sobel test of mediation}
\usage{
calc_sobel(x, y, z, y.names, z.names)
}
\arguments{
\item{x}{binary numeric vector of epi-DNA states}

\item{y}{matrix of cis gene expression ncol(x) must be equal to length(x)}

\item{z}{matrix of trans gene expression ncol(z) must be equal to length(x)}

\item{y.names}{names of cis genes in y}

\item{z.names}{names of trans genes in z}
}
\description{
Performs sobel test of mediation
}
