% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_data.R
\name{cls2df}
\alias{cls2df}
\title{\code{cls2df} reads from .cls file into data frame}
\usage{
cls2df(f)
}
\arguments{
\item{f}{.cls file name}
}
\description{
\code{cls2df} reads from .cls file into data frame
}

