% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iEDGE_util.R
\name{construct_iEDGE}
\alias{construct_iEDGE}
\title{Construct iEDGE input object}
\usage{
construct_iEDGE(cn, gep, cisgenes, cn.fdat, gep.fdat, cn.pdat = NA,
  gep.pdat = NA)
}
\arguments{
\item{cn}{matrix of epi-DNA alterations (m rows of alterations by n samples)}

\item{gep}{matrix of gene expression (g rows of genes by n samples)}

\item{cisgenes}{list of cis genes corresponding to each row in cn}

\item{cn.fdat}{data frame of row annotations for cn, must have column of row identifiers, and optionally column of direction identifiers if one-sided DE tests are applied}

\item{gep.fdat}{data frame of row annotaions for gep, must have column of row identifiers}

\item{cn.pdat}{optional data frame of column annotations for cn}

\item{gep.pdat}{optional data frame of column annotations for gep}
}
\description{
Construct iEDGE input object
}
