% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gistic2ge.R
\name{do_gistic2ge}
\alias{do_gistic2ge}
\title{\code{do_gistic2ge} wrapper for performing differential expression test}
\usage{
do_gistic2ge(ge, gistic, case = c(1, 2), control = (0),
  interaction_type = "cis", test = "limma", alternative = "equal",
  boundary = "peak", logged = TRUE)
}
\arguments{
\item{ge}{internal ge object}

\item{gistic}{internal gistic object}

\item{case}{values to be considered altered samples in gistic lesions file, default c(1,2)}

\item{control}{values to be considered unaltered samples in gstic lesions file, default c(0)}

\item{interaction_type}{"cis" or "trans" type of analysis to performed}

\item{test}{"limma" or "ttest" DE test to be used, default "limma"}

\item{alternative}{"equal", "greater", or "less" is DE test one-sided?
rule of thumb for trans alternative = "greater", for cis, alterative = "greater" for amp and "less" for del}

\item{logged}{TRUE if gene expression is log2-transformed, DE is performed on log2 expression values}
}
\value{
data frame with DE results
}
\description{
\code{do_gistic2ge} wrapper for performing differential expression test
}

