% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gistic2ge.R
\name{do_gistic2ge_all}
\alias{do_gistic2ge_all}
\title{\code{do_gistic2ge_all} wrapper for \code{do_gistic2ge} for performing multiple differential expression tests}
\usage{
do_gistic2ge_all(gistic_ge_amp, gistic_ge_del, logged = FALSE)
}
\arguments{
\item{gistic_ge_amp}{internal gistic_ge object for amplifications}

\item{logged}{TRUE if gene expression is log2-transformed, DE is performed on log2 expression values}
}
\value{
list of data frames with DE results
}
\description{
\code{do_gistic2ge_all} wrapper for \code{do_gistic2ge} for performing multiple differential expression tests
}

