% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gistic2ge.R
\name{do_test_limma}
\alias{do_test_limma}
\title{\code{do_test_limma} performs limma differential expression test}
\usage{
do_test_limma(mat, design, alternative = "equal", logged = TRUE)
}
\arguments{
\item{mat}{gene expression matrix}

\item{design}{design matrix}

\item{alternative}{must be one of "greater", "less" or "equal", P.Value and adj.P.Val turned to one-sided results if specified}

\item{logged}{TRUE if gene expression is log2-transformed, limma model is performed on log2 expression values}
}
\value{
limma topTable data frame
}
\description{
\code{do_test_limma} performs limma differential expression test
}

