% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{iEDGE_combine}
\alias{iEDGE_combine}
\title{Combines alteration data (e.g. GISTIC2) and gene expression data, using a mapping data frame for matching samples}
\usage{
iEDGE_combine(alt, gep, mapping, mapping.cn = "CN", mapping.gep = "GEP",
  uppercase = TRUE)
}
\description{
Combines alteration data (e.g. GISTIC2) and gene expression data, using a mapping data frame for matching samples
}
