% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iEDGE_util.R
\name{iEDGE_DE}
\alias{iEDGE_DE}
\title{iEDGE_DE performs differential expression analysis and pathway enrichment and saves to text tables}
\usage{
iEDGE_DE(cn, gep, cisgenes, header, gepid, cnid, f.dir.out, gs.file,
  gs.file.name, fdr.cis.cutoff = 0.25, fdr.trans.cutoff = 0.05,
  min.group = 3, min.drawsize = 3, cndir = "alteration_direction",
  cis.onesided = TRUE, trans.onesided = FALSE, ...)
}
\description{
iEDGE_DE performs differential expression analysis and pathway enrichment and saves to text tables
}

