% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iEDGE_UI.R
\name{iEDGE_UI}
\alias{iEDGE_UI}
\title{iEDGE_UI makes the user interface for iEDGE reports}
\usage{
iEDGE_UI(cistab, cisfulltab, transtab, cn, gep, cisgenes, outdir,
  jsdir = file.path(path.package("iEDGE"), "javascript"), cmijsdir, cmi,
  altid = "Unique.Name", geneid = "accession")
}
\description{
iEDGE_UI makes the user interface for iEDGE reports
}

