% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_data_lymphoma2010.R, R/util.R
\name{main}
\alias{main}
\title{\code{main} example for running Gistic2GE pipeline}
\usage{
main()

main()
}
\description{
\code{main} example for running Gistic2GE pipeline

\code{main} example for running Gistic2GE pipeline
}

