% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_GISTIC2.R
\name{make_GISTIC2_focal}
\alias{make_GISTIC2_focal}
\title{make_GISTIC2 preprocess GISTIC2 data}
\usage{
make_GISTIC2_focal(gistic_in, all_genes, qvalue_focal, all_lesions = NA,
  f.amp = NA, f.del = NA)
}
\description{
make_GISTIC2 preprocess GISTIC2 data
}
