% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_data.R
\name{match_samples}
\alias{match_samples}
\title{\code{match_samples} match samples found in ge and gistic objects}
\usage{
match_samples(gistic, ge, fmap.file, fmap.gistic.col = "snp.array.ID",
  fmap.ge.col = "mrna.array.ID", ge.confounder.file = NA)
}
\arguments{
\item{gistic}{gistic object}

\item{ge}{ge object}

\item{fmap.file}{file containing mapping from ge to gistic samples}

\item{fmap.gistic.col}{column name in fmap.file containing gistic ids}

\item{fmap.ge.col}{column name in fmap.file containing ge ids}

\item{ge.confounder}{sample confounder indicator file name, default = NA, no confounder}
}
\description{
\code{match_samples} match samples found in ge and gistic objects
}

