% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{\code{process_data} process gistic and ge data from text files}
\usage{
process_data(amp.peak.file, amp.region.file = NA, del.peak.file,
  del.region.file = NA, lesions.file, ge.file, fmap.file,
  ge.confounder.file = NA, binary = TRUE, match.by = c("Descriptor",
  "cytoband"), fmap.gistic.col = "snp.array.ID",
  fmap.ge.col = "mrna.array.ID", row.ind = FALSE)
}
\arguments{
\item{amp.peak.file}{GISTIC amp genes in peak file}

\item{amp.region.file}{GISTIC amp genes in region file}

\item{del.peak.file}{GISTIC del genes in peak file}

\item{del.region.file}{GISTIC del genes in region file}

\item{lesions.file}{GISTIC all_lesions file}

\item{ge.file}{Gene Expression res file}

\item{fmap.file}{file mapping tabular file from GISTIC to GE sample names}

\item{ge.confounder.file}{confounder indicator cls file, must match order in fmap.file}

\item{...}{see addition arguments passed to \code{match_samples, read_res, read_cis_genes, read_gistic}}
}
\value{
list containing :
\item{amp}{amplification gistic and ge}
\item{del}{deletion gistic and ge}
}
\description{
\code{process_data} process gistic and ge data from text files
}

