% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_cis_genes}
\alias{read_cis_genes}
\title{\code{read_cis_genes} reads cis genes from GISTIC files, helper function for read_gistic}
\usage{
read_cis_genes(file.in.peak, file.in.region, cn.dir)
}
\arguments{
\item{file.in.peak}{is the file containing in_peak genes}

\item{file.in.region}{is the file containing in_region genes}

\item{cn.dir}{must be one of "amp" - Amplification or "del" - Deletion}
}
\description{
\code{read_cis_genes} reads cis genes from GISTIC files, helper function for read_gistic
}

