% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_gistic}
\alias{read_gistic}
\title{\code{read_gistic} reads from GISTIC files}
\usage{
read_gistic(lesions.dir, genes.peak.dir, genes.region.dir, binary = T,
  match.by = c("Descriptor", "cytoband"), cn.dir)
}
\arguments{
\item{lesions.dir}{is the file containing all_lesions}

\item{genes.peak.dir}{is the file containing in_peak_conf## genes}

\item{genes.region.dir}{is the file containing in_region_conf## genes}

\item{binary}{= T, if TRUE (default) get binary indicator rows for sample alteration status, otherwise actual copy change reported}

\item{match.by}{c(first, second), first = unique id column name in lesions.dir (default = Descriptor), second = unique id column name in genes.peak.dir (default = cytoband)}

\item{cn.dir}{must be one of "amp" - Amplification or "del" - Deletion}
}
\description{
\code{read_gistic} reads from GISTIC files
}

