% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_gmt.R
\name{read_gmt}
\alias{read_gmt}
\title{\code{read_gmt} read from gmt file}
\usage{
read_gmt(f, split.char = "\\t")
}
\arguments{
\item{f}{name of gmt file}

\item{split.char}{delimiter}
}
\description{
\code{read_gmt} read from gmt file
}

