% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_res}
\alias{read_res}
\title{\code{read_res} converts .res data into internal gene expression data format}
\usage{
read_res(f, row.ind = FALSE)
}
\arguments{
\item{f}{gene expression .res file}

\item{row.ind}{whether first row has numeric row indices, default = FALSE}
}
\description{
\code{read_res} converts .res data into internal gene expression data format
}

