% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hyperEnrichment.R
\name{run_hyperEnrichment}
\alias{run_hyperEnrichment}
\title{\code{run_hyperEnrichment} performs hyperenrichment test from gistic2ge DE results on gene sets}
\usage{
run_hyperEnrichment(drawn, categories, ntotal, min.drawsize = 4, mht = TRUE,
  verbose = FALSE, order = TRUE, hypercol = "fdr", hyperthres = NULL)
}
\arguments{
\item{drawn}{list of genesets, usually from user-defined queries}

\item{categories}{list of genesets, usually from pathway databases}

\item{ntotal}{total number of genes}
}
\value{
limma topTable data frame
}
\description{
\code{run_hyperEnrichment} performs hyperenrichment test from gistic2ge DE results on gene sets
}
