% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hyperEnrichment.R
\name{run_hyperEnrichment_wrapper}
\alias{run_hyperEnrichment_wrapper}
\title{\code{run_hyperEnrichment_wrapper} wrapper for run_hyperEnrichment}
\usage{
run_hyperEnrichment_wrapper(gs.tab, ntot, gs.pathway)
}
\arguments{
\item{gs.tab}{DE results table from gistic2ge}

\item{ntot}{total genes}

\item{gs.pathway}{list of genesets}
}
\value{
list of data frames of hyperenrichment results
}
\description{
\code{run_hyperEnrichment_wrapper} wrapper for run_hyperEnrichment
}
