% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iEDGE_util.R
\name{run_iEDGE}
\alias{run_iEDGE}
\title{run_iEDGE is the main wrapper for iEDGE, assumes processed data}
\usage{
run_iEDGE(dat, header, outdir, gs.file = NA, gepid = "SYMBOL",
  cnid = "Unique.Name", cndir = "alteration_direction",
  fdr.cis.cutoff = 0.25, fdr.trans.cutoff = 0.05, fc.cis = NA,
  fc.trans = NA, min.drawsize = 3, onesided.cis = TRUE,
  onesided.trans = FALSE, uptest = "Amplification", downtest = "Deletion",
  gs.file.name = "c2.cp.v5.0", min.group = 2, mutinfo.seed = 7,
  mutinfo.nsamples = 500, mutinfo.bins = 5, prune.method = "cmi",
  prune.col = "pvalue", prune.thres = 0.05, hyperthres = 0.25,
  cis.boxplot = TRUE, trans.boxplot = TRUE, bipartite = TRUE)
}
\arguments{
\item{outdir}{output directory}

\item{iEDGE}{datalist consisting of cn (alteration), gep (gene expression), and cisgenes (list of genes in each alt)}
}
\description{
run_iEDGE is the main wrapper for iEDGE, assumes processed data
}

