% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_data_lymphoma2010.R, R/util.R
\name{run_process_data}
\alias{run_process_data}
\title{\code{run_process_data} example for processing gistic and ge files into internal gistic_ge objects}
\usage{
run_process_data()

run_process_data()
}
\description{
\code{run_process_data} example for processing gistic and ge files into internal gistic_ge objects

\code{run_process_data} example for processing gistic and ge files into internal gistic_ge objects
}

