% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.save.R
\name{util.save.rds}
\alias{util.save.rds}
\title{\code{util.save.rds} save to .RDS, and optionally .xls (for list of data frames)}
\usage{
util.save.rds(x, f.dir = NA, f.header = "ricover2010_", f.name)
}
\arguments{
\item{x}{R object to save}

\item{f.dir}{directory to save to, default = NA, current working directory}

\item{f.header}{file header}

\item{f.name}{file name (appended to file header)}
}
\description{
\code{util.save.rds} save to .RDS, and optionally .xls (for list of data frames)
}

