% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.save.R
\name{util.save.xlsx}
\alias{util.save.xlsx}
\title{\code{util.save.xlsx} save to .xlsx}
\usage{
util.save.xlsx(x, f.dir = NA, f.header = "ricover2010_", f.name)
}
\arguments{
\item{x}{R object to save, must be a named list of data.frames, row limit =65535 for each data frame}

\item{f.dir}{directory to save to, default = NA, current working directory}

\item{f.header}{file header}

\item{f.name}{file name (appended to file header)}
}
\description{
\code{util.save.xlsx} save to .xlsx
}

